Initial work for task add and refactor
